+++ src/fmindex-collection/rankvector/MultiBitvector.h
@@ -12,9 +12,9 @@
 
 namespace fmindex_collection::rankvector {
 
-template <uint64_t TSigma, BitVector_c Bitvector = ::fmindex_collection::bitvector::Bitvector>
+template <size_t TSigma, BitVector_c Bitvector = ::fmindex_collection::bitvector::Bitvector>
 struct MultiBitvector {
-    static constexpr uint64_t Sigma = TSigma;
+    static constexpr size_t Sigma = TSigma;
 
     std::array<Bitvector, TSigma> bitvectors{};
